HTTP/2 feature parity, TLS hardening, and endpoint extraction
- Extract shared endpoint logic into endpoints.c/h (health, metrics body
  generators, latency/status/method counter updates, hex validation,
  ACME challenge handler for HTTP/2)
- HTTP/2 endpoint parity: /health, /ready, /alive, /metrics, ACME
  challenges, static files all serve identical content over h2
- Per-stream request IDs and response tracking (X-Request-ID, Cache-Control)
- Unified h2_process_stream_request() routing for HEADERS+DATA END_STREAM
- nghttp2 CVE protections: Rapid Reset (CVE-2023-44487), CONTINUATION
  flood (CVE-2024-28182), SETTINGS/ping flood limits
- TLS hardening: Mozilla Intermediate cipher list, NO_RENEGOTIATION,
  defense-in-depth flags, consistent init+reload
- Remove dead code: connection_set_callbacks(), h2_send_error()
- H2 stream tier downgrade after request processing
